+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "bcrypt": "^5.1.1",
     "bcryptjs": "^2.4.3",
     "parse": "^5.3.0",
     "react": "^18.3.1",
