+++ src/Components/Navbar/DesktopNavigation.js
@@ -1,22 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import NavlinksAfterLogin from "./NavlinksAfterLogin.js";
 import NavlinksBeforeLogin from "./NavlinksBeforeLogin.js";
 import "./Navbar.css";
 import { Link, useLocation } from "react-router-dom";
 
-function DesktopNavigation() {
+function DesktopNavigation(authenticatedUser) {
   const location = useLocation();
-  const [authenticatedUser, setAuthenticatedUser] = useState(false); // assuming false means not authenticated
-
-  // Check if the user is authenticated or based on path
-  useEffect(() => {
-    // You can update this based on your auth logic
-    if (location.pathname === "/") {
-      setAuthenticatedUser(false);
-    } else {
-      setAuthenticatedUser(true);
-    }
-  }, [location]);
 
   // Function to handle logo click
   const handleLogoClick = (e) => {
@@ -34,7 +23,7 @@ function DesktopNavigation() {
             <img src="images/icon.png" alt="logo" width="100" height="90" />
           </Link>
         </div>
-        {authenticatedUser ? <NavlinksAfterLogin /> : <NavlinksBeforeLogin />}
+        {authenticatedUser ? <NavlinksBeforeLogin /> : <NavlinksAfterLogin />}
       </nav>
     </div>
   );
