+++ src/Components/Navbar/NavlinksBeforeLogin.js
@@ -7,7 +7,10 @@ function Navlinks({ isClicked, closeMenu }) {
     <nav className="navlinks">
       <ul>
         <li onClick={() => isClicked && closeMenu()}>
-          <Link to="/">Log in</Link>
+          <Link to="/loginTherapistUser">Therapist Log in</Link>
+        </li>
+        <li onClick={() => isClicked && closeMenu()}>
+          <Link to="/loginPatientUser">Patient Log in</Link>
         </li>
       </ul>
     </nav>
